Extract auth methods
